[bin] clean up install script, create linking script
I found that there was a definite use case where I wanted to
do the linking without running any of the install stuff.

I decided to pull out the shared functions and create a linking
script and an install script. Modularization is nice because I'll
use the messages script if/when I get around to writing an uninstall
script.