vim: goodbye pathogen, hello native packages
I've known about Vim 8's packages for a while as a native solution to
package management, but I've never gotten around to reading about it
enough to transition off of pathogen.

I finally got the motivation, and it was easier than I thought! Most of
the effort went to deciding which packages to be start/opt (which I
really like, actually), and writing a home-brewed solution to Helptags.