vim: update .gitignore, delete .gitkeep files
I had been using .gitkeep files to hold down the pack/mine and
pack/plugins directories, but now that there are submodules, this isn't
necessary.

Also, it seems that the best way to go about submodules is still
gitignore the whole directory (as I add plugins manually based on the
development at hand (and may not want some plugins as part of my install
everywhere)). Note that this means submodules need to be added with
'-f'.