[bash] add git aliases, append history per command
One pain point with history was that if I opened a new split in tmux (or
a new term) the history isn't available yet. By adding "history -a" to
the PROMPT_COMMAND, it will append after each item, being instantly
available to new terms.

This commit also adds a few git aliases along with an alias to reload my
bashrc (as that was something I was doing often enough).