[bash] make source_if_exists function always return 0
I have a "source_if_exists" function which was really just to enhance
readability, make changes easier, etc.

However, it was returning 1 if the file did not exist, and if that call
was the last in the bashrc, it would display the [1] in my bash prompt.

This commit changes the function to be an if statement instead of a &&,
thus returning 0 if the file does not exist.