[cleanup] move all top level dotfiles to subdirs
Previously, I would have a directory for tools with multiple dotfiles
(such as bash) and would just house the dotfiles that had a 1-1 ratio of
tool-dotfile (i.e. tmux.conf) in the root of this repo.

I decided to move to a 1-1 on tool-subdir, even if there is only one
file. This allows symlinking that whole directory to ~/.config should it
be that style of dotfile and it allows for a cleaner repo overall (files
on the top are not dotfiles, distinguishing between things like the
Dockerfile, etc).

I may also start adding notes on tools in README's for the respective
subdir. We shall see.

I also moved neovim->nvim such that the binary name is the name of the
subdir.