[doc] update README with safer/more accurate install instructions
Previously I simply recommended a curl-into-bash of the installer. I
have since come to realize this isn't the safest option (and respect
that I shouldn't recommend it). As such, I've updated the README to
encourage the user reads the installer.

Additionally this documents the link.sh script, which is really the
workhorse of the installer (in the sense that it alone sets up the
dotfiles, the installer just installs them and other things I want).