[tmux] pipe copy-selection to xclip for copying
After noticing strange behavior in tmux with copy/pasting, I found the
root issue (described in github issue tmux#1119):

set clipboard on (the default) will try to set the system clipboard by
sending it an escape sequence. The terminal emulator I am currently
using (st) *can* handle this escape sequence, but the buffer size it has
for it is abysmally small (1024 chars).

Thankfully, the workaround is rather simple: disable this escape
sequence sending (to avoid a race condition with xclip) and change
the binding of 'y' in copy-mode-vi to pipe to xclip instead (which can
handle a much larger buffer).

This will obviously not hold up on macOS, but for now I am working only
on Linux so this isn't a big issue.