[mutt] adding bare-bones mutt config
I was using Luke Smith's mutt-wizard for a while, which was great, but I
reached the point where I wanted to really understand what was going on
in the tool.

So I scrapped it all and started from scratch, reading the manual and
tailoring the level of complexity to my not so intense needs.

Colors are lacking, but otherwise this is a pretty functional little
mail client!