[mutt] add muttrc.local concept, update mailcap, bindings
Given that I am not *always* using my karleco account, I added the
concept of sourcing a "muttrc.local" which is not checked in. This file
can be all the account info, or just a file that sources the
accounts/karleco.muttrc.

Small changes to muttrc including some key bindings, mailcap for images,
and adding a BCC to myself on send so it shows up in =INBOX.