[mutt] use `abook` for contact management
The primary functional change in this commit is that it binds <tab> in
the editor to query `abook`, a separate program for contact managent.

Further, it adds a macro "ga" to add the current sender to abook.

Other than that, changes tab/up/down to go by unread instead of new and
changes around whitespace/organizes bindings better.