[bash] don't set TERM in prompt
When I first created this prompt, it was heavily inspired by
@mathiasbynens. So, when he set TERM in the beginning, so did I.

However, I found the other day that this *can* have unintended
consequences. In particular, in a Linux Virtual Console (TTY, no X
server), the xterm... TERM can throw off tmux and not cause it to clear
the screen correctly (as tmux thinks that the thing connected to the
client is an xterm, but it's not).

In this case, we need TERM to be linux, but *really*, I should have
never been mucking around in this space anyways.

It seems that TERM is linux by default (works), and tmux sets
tmux-256color, so really it all works out if the prompt doesn't try to
do anything smart :)