[tmux] add slstatus to statusline, other theme tweaks
Most changes are just reorganizing the config to be a bit cleaner.

The clever part to this (not really, but I thought it was cool) is that
if a command for the statusline doesn't terminate, tmux just grabs the
topmost line. slstatus with the -s flag does just this--printing to
STDOUT every second (at which point tmux will grab the latest and
greatest).