tmux: remove explicit setting of $TERM
This was a couple lines of config that I pasted in a long time ago when
I was desperately trying to get italics to work in iTerm2.

I am now using `st` as my true color terminal emulator, and it doesn't
need this escape code hack. I have given up on italics (not worth the
time or effort), so here's to simpler configs!

More importantly though, this setting of $TERM to be 256 colors really
messes up Vim in the Virtual Console, because it thinks it is 256 colors
when it is really 8 (16 with bold which renders bright), and
colorschemes get all funky. Let tmux set its own TERM value, thank you!