bash: prefer Vim over NeoVim in Virtual Console
If $DISPLAY isn't set (we aren't in X), I have come to prefer Vim over
NeoVim.

It pains me, because I *feel* like NeoVim should work fine, but for the
life of me I cannot get 256 colors (or even linux-16color) to render
properly in the VC.

The problem with NeoVim is that it took out the concept of t_XX
settings, which means that t_Co is perpetually at 256 colors which means
that colorschemes will try to use colors outside of the 8 base colors
which makes for some wonky colorschemes (try visual mode without any
indication of the selection... fun stuff).

Vim still handles t_Co=8 well, and correctly identifies it in both
TERM=linux and TERM=screen/tmux, so I've opted to prefer Vim in in the
VC.