tmux: set TERM explicitly, but be smart
A few commits back I took out the setting of TERM in tmux. I was under
the impression this wouldn't be an issue (for some reason, likely a
non-dead server, I thought tmux was being smart about the client and
setting the TERM properly).

However, st was connecting and getting TERM=screen, which only has 8
colors.

This commit adds a conditional to set the default-term properly for both
the VC and st.