vim/colors: add colorscheme for the Linux Virtual Console
I've been playing around a lot recently in the Linux Virtual Console (no
X server). Having Vim colors (at the bare minimum for functional stuff
like visual mode / search highlight) is important.

I tried to get gruvbox to work, but in the end I needed to build off
something much simpler that made no assumptions about how many colors
were available--so, this 8 color theme builds off of the "default"
theme, which Vim detects properly to be 8 colors if in the VC and
everything should "just work".

Note that NeoVim does not acknowledge t_Co=8 mode, so this theme will
likely break in NeoVim in the VC (but should be fine in a X terminal
emulator).