readline: use standard tab completion
This is an experimental shakeup. I've spent a year with this vim-style
tab completion (in fact it was the first and only setting in my original
inputrc).

However, I find that bash has high-stakes scenarios where you *really*
don't want to tab-complete the wrong thing. So while this is usually
useful, I want to dial back my trigger happiness with the TAB key :)

May revert later...