mutt: add macro for pandoc HTML email generation
Using neomutt's multipart/alternative grouping mechanism, we can
generate an HTML email (with a custom stylesheet!) and then group it
with the markdown plaintext used to generate it.

This is a nice compromise between those who like to view the HTML and
those who prefer plaintext.

One open question is whether this holds up well in terms of preserving
the HTML of the thread. Only time will tell!

NOTE: requires neomutt . As of now, no multipart/alternative support for
mutt.