cleanup: remove Dockerfile
I've always played with the idea of automating the installation of this
repo. First there was a Vagrantfile, then a Dockerfile, and now I just
broke down and wrote a basic symlink manager specifically for dotfiles.

With dedicating the installation / management to this tool, I no longer
maintain or use the Docker installation. This commit removes it
accordingly.