mutt: update karleco account to use Maildir
This uses the offlineimap config that I set up in a previous commit.

The paradox of the UNIX philosophy -- I eventually only want one tool
that does each job well (offlineimap for syncing, mutt for viewing,
msmtp for sending...). But when I am a novice user, I want one tool that
does it all for me for the sake of getting started.

Or maybe this is because I'm coming from the world of webmail where it
was clicks away from creating an account to sending emails?