cleanup: remove alacritty and italic terminfo files
I stopped using alacritty when I moved off of my mac. The config had
nothing special to it other than the size of the font / the gruvbox
colors, so I'm removing it for now (beauty of git, can always go back to
it!). I'm happy with `st` on Linux.

The terminfo files were from _way_ back when I was an appearance-driven
vimmer and looking to get italics in the terminal. They turned out to be
way more of a hassle than I'd like (esp. when combo'd with tmux, which
is currently crucial to my scrollbackless-st workflow), so I haven't
used the files in years and don't see myself going back to the italics
in terminal life.