vim: fix dirvish vs netrw decision
There is an outstanding bug in... something, that makes dirvish and
netrw not play nicely together. I was previously checking for the
presence of the "vim-dirvish" directory to see if we should disable
netrw.

This breaks down because it will always be there, even if the submodule
is not 'init' and 'updated'.

Changing the check for vim-dirvish/plugin correctly represents if we are
actually using dirvish.