install: don't use custom dotfile manager
I found that I was setting up my dotfiles on quite a few new accounts
recently (mostly due to exploring different OSs). When I get a new
machine / account, I want to be able to install the basics (and a few
extras) without needing the dependency on my dotfile manager (or even
git or bash).