sh: simpler PS1, migrate w3m funcs to common aliases
I'm experimenting with a one-line PS1. It's been a _long_ time
since I've used such a simple prompt, but I want to give it a
go. It's an interesting exercise in simplicity -- if the paths
are so long that they get in the way of the prompt, maybe there's
a bigger issue?

Also `git log` on OpenBSD with a 2 line prompt pushes back the
first line out of sight, so that's a bummer.