sh: Add HISTFILE, simplify prompt
In ksh for OpenBSD, we need to set HISTFILE so that history is saved.
However, we don't want to clobber HISTFILE if already set (i.e. by bash)

This commit also simplifies the prompt significantly. I've been playing
around with a no-path prompt and grown quite fond of it. This is the
formula I've settled on for now.