bin: Add psync, a tool to sync pash passwords
pash is a simple password manager that stores each password as a GPG
encrypted key (see submodules for link). I like to backup to a central,
always on, host in my apartment (defaulting to a raspberry pi) so that
I can easily transfer passwords between my desktop and laptop.

There's a slight chance of password loss with the "sync" action (i.e.
someone else updated a password that you have a stale copy of), so it
defaults to needing to specify push/pull/sync for safety.