install: Trim down what's installed by default
There's a bunch of things here that I _don't_ necessarily want on every
machine I put my dotfiles on. For example, offlineimap or mbsync.

I took out gdb too because I don't use it often and I'm happy to symlink
it manually on the few machines that I do use it on.