cleanup: Remove offlineimap config
I don't frequently delete old dotfiles, on the offchance I'm on a
machine where I can't install my preferred tool and have to resort to a
backup (and look, hey! I've got a config for that).

screen is a great example of this -- never used, really, but not
inconceivable that it might be of use one day where tmux isn't
available.

offlineimap is not one of those "maybe I'll need this" cases.