git: Update config to set pull default behavior
It seems newer versions of git print a verbose warning if you type
"git pull" and don't have a default set.

In general, I want it to error if it can't fast forward (then I'll
run it again with rebase). I almost never want it to merge, but I
like to know that I had to rebase in the first place.