gtk: Add gtk-3.0 settings file for OpenBSD firefox
The OpenBSD build of firefox uses emacs readline style bindings by
default, which gets realllllllly frustrating when you're doing something
like trying to close the tab and it's trying to delete a word :)

While this is a 2 line file, I'm checking it in because the effort I
went through to find the proper fix was > 5 minutes of time, and I'm
most certainly not going to remember it.

It lives under ~/.config/gtk-3.0 . Discovered on this great blog post:

    https://www.c0ffee.net/blog/openbsd-on-a-laptop/

Heresy indeed!