git: Change contact git@ -> alex@alexkarle.com
After sleeping on it, I decided it feels more natural to email
USER@domain instead of FUNCTION@domain.

I know there's a lot of companies out there that do the latter (i.e.
support@company.com), but I'm a single person, and I want my commits to
represent that (and not give the illusion of being an entity).