sh: Update to use include() function
The exit status of the pipeline:

    [ -r foo ] && . foo

is false if foo doesn't exist. This then becomes the exit status of
sourcing my config, which then becomes displayed in a new login prompt.

The solution? Use a true if block (which will always return true). Wrap
it in a function to keep it clean.