sh: Change default prompt to include cwd
I've been using a prompt like so for a while:

[user@host]$

Which is nice because it's consistently short, helpful when I'm
SSH-ing around, and nicely minimal.

But after testing ash as a login shell, I grew fond of the simple
"\w \$" prompt:

~/dotfiles $

So I'm going back to the cwd-in-prompt crew (admittedly, I mostly
use my fancy bash prompt which is multi-line--this is the prompt
I use on OpenBSD and fresh Linux installs).

I confirmed \w is portable to pdksh, bash, and busybox ash. dash
doesn't recognize it, but that's OK (can't imagine using dash as
a login shell...)