Put the dot back in dotfile
I've been feeling for a while now that all these directories with one
file are just confusing (i.e. mbsync/mbsyncrc ...) and so I've resolved
to "put the dot back in dotfile" -- remove the nested subtrees, and
mirror a real $HOME directory. Anything that I don't really want to be
under $HOME, I'm tucking away in ~/etc, which feels appropriate (i.e. vc
configs or perl setup, etc).

This approach was inspired by this HN comment [0].

The beauty in this is that you can then use git to inspect the
how $HOME looks as a worktree of this repository. The key thing here is
that **it's not a full work tree!** there's no .git file in $HOME that
links the two, so a rogue "git status" in $HOME won't trigger any
output. All changes go through the "cfg" alias which diffs the repo
(~/.cfg, bare) and the worktree ($HOME).

On machines where I only want a fraction of the dotfiles (i.e. work), I
can still do a full checkout and manually symlink.

This was super easy to change with the exception of the submodules (hot
dang were they tricky to move correctly) -- note to future self, the
clone URL is in .git/config (not just .gitmodules)...

Not much functional change, except for removing my zshrc, which I
haven't used in ages anyways (oh my zsh on macOS? those were the early
days for sure :) ). From a git perspective, there are no more
.gitignores, being replaced with a single '*' entry. Despite it not
being a true worktree I want to be able to `git add .` (force of habit)
and not add my whole $HOME.

I'll be curious to see how this simplifies/complicates working on
multiple machines. I don't plan to maintain multiple branches, but that
could change.

[0]: https://news.ycombinator.com/item?id=11071754