ed: Disable rlwrap warnings, logging
On OpenBSD rlwrap prints a warning on startup that it
cannot clomplete filenames outside the cwd. This is OK
(likely an unveil call), but should be suppressed.

Likewise, the logfile option was initially added because
I was noticing strange behavior with rlwrap, but I haven't
seen it recently (and don't want to start building cruft
in /tmp).