vi: Remove tab expansion / auto wrapping
While my original use case for vi(1) on OpenBSD was commit messages
(and I still do use it for that), I've grown to use it for quick
code-edits too; and in these cases, having the tab expand and lines
wrap have been a real hassle.

The config file is still useful -- having the ruler is helpful for
writing text that should be wrapped (i.e. mailing list posts, commit
message) -- but it's now way less intrusive.