install: Change from git-worktree to symlink farm
We've been here before... about 5mo ago, I decided to give the "just
checkout into $HOME" a shot with a dedicated "cfg" script to push/pull
etc. This worked OK, but it had a couple flaws:

  1. Ignoring everything was annoying (needed git add -f)
  2. The bare repo in ~/.cfg doesn't track other remotes
     (there was no way to see ahead/behind origin)
  3. It promotes an all or nothing approach / relied on DOT.local
     for site-local customizations

This patch goes back to a more traditional style of symlink-farm, with
a twist! The twist is that the .{bash_,}profile files are NOT in git,
as they only contain one-two lines worth versioning. This allows us
to generate them on install and put site-local modifications there.

In addition, since I noticed I was only really using git/tmux/vi/sh on
my remote servers, I ended up splitting the install into skel (the bare
necessities) and install.sh. I expect to use install.sh on the desktop
and skel on the server, and the `lnkdot` script can easily add more
to an existing installation.

Finally, note that the generated profiles put $DOTFILES/bin on the path
instead of copying it to ~/bin. This allows easier separation of what
is and what isn't version controlled out of my personal scripts. `e`
has been updated accordingly to find it's inputrc relative to itself.