irc: Add tic(1) -- ic(1) in tmux(1)
This is a fun little extension of ic(1). I found it really annoying to
have what I was typing double-shown to me (i.e. once I sent the message,
it'd appear multiplexed with the stdin feed). One solution I found was
to open a completely separate viewing panel in tmux(1) that would be
tailing the log file, with only a 1 line prompt-panel for ic(1) (with
its stdout redirected to /dev/null).

This works really well, but was a pain to set up--tmux commands to the
rescue! I was able to automate the panel setup in a new session, which
is _so_ cool :)