xterm: Change default colors & fix font size
One of the OpenBSD updates updated the DPI in a way that made it so that
things are bigger [1]. Right now, I'm fixing it on an individual
application basis rather than forcing the DPI, but I might change it
later!

Also, homebrew colors are harder on the eyes than lightgrey!

[1]: https://marc.info/?l=openbsd-tech&m=163674121630769&w=2