Add git-grep alias and git-grep-editor tools
This is a fun exploration of UNIX as an IDE. The problem: I want
to be able to quickly jump to occurrences as specified in git-grep.
The solution? I either need to embed git-grep in an IDE (vim), or
I need to write a tool that parses git-grep and knows HOW to open
an editor.

This patch adds such a tool, to stay true to the UNIX philosophy,
warts and all. The tool has two smaller bits--one to parse the
"file:number:desc" and launch the editor, and another to parse
git-grep.