bin: Add `e` program to launch ed(1) with a prompt
I'm back on the ed(1) train, and the prompt is definitely a must.
Pro tip of the day--I always wanted to filter lines through an
external program and read it back, and thought it wasn't possible
in one command. Something like vi's !motion command.

Turns out it is! The trick is that the `e` command can take in a
!command, *replacing the contents of the buffer without changing
the default filename*. So, to reformat the buffer, just write it
out first and then:

	> e !fmt < %

Voila! I used it on this commit :)