X11: Add primitive .xsession for xenodm(8)
I know I'm leaving a bit of config behind here (specifically
the "scroll with the red thinkpad nub" wsmouse calls that I found
way back on reddit), but to stay true to my license, this is 100%
clean-room reimplemented.

And hey, I like the color (ok the color is from the OpenBSD website
dark theme... but that's OK to lift, I think?).