cwm: Add basic cwm(1) config
This is a great example of where redoing the config has left
me (1) in awe of all the features I haven't used / don't need,
and (2) happier with where I landed.

The thing I like best in this version is that I explicitly
unbound all, which means that I won't accidentally have keys
get captured while doing other random tasks!

There's definitely a few things missing but it's check-in-able!