X11: Add xinput config for trackpoint scrolling
This was a fun one to recreate. I knew roughly what I needed from
some googling, but no one seemed to agree and no one linked any doc.

I was able to find the doc by searching for "WS Pointer" in the xenocara
repo and sure enough -- voila! doc!

I tried to document what I learned--finding the button id's was a mix
of Arch wiki and trial and error.