bin: Add new `demo(1)` for `script(1)` support
I frequently want to get the output of commands to a coworker
and script(1) would be a nice tool to use, with a few caveats:

1. Escape sequences don't transfer well
2. Muscle memory has me using my aliases

demo(1) solves both of these by spawning a new shell without
my aliases and with TERM=dumb to prevent the sequences. It also
sets a blank prompt, which is frequently desirable when sharing
snippets (frequently my PWD is not useful / adds noise).

Excited to try it out tomorrow!