nvim: Add initial from-scratch neovim config
With acme inspired colors, of course!

I've held off on adding full editor support to this repo for a
while. I made it really far on vi(1) alone :) However, as I get
closer to using this as a daily driver, I need a _slightly_ more
powerful editor!

I chose neovim because:

  1. I didn't need to finagle t_Co & friends to get the right colors
  2. I think vim9's new scripting language should have been lua

So in protest of a new VimL, we have a full-lua config! We'll see
how far it goes.

(I've also been mucking around with acme(1) itself, but I don't want
to be that attached to my mouse :/)