tmux: Add plumb command and key-binding
Inspired by acme(1), I really wanted a way to take the
tmux selection contents and open them in the appropriate
application.

I spent a while exploring st(1)'s patch for this, but I didn't
really want to use the vte.sh pre-prompt updating of a cwd
and the "simple" patch only supports execution from st's cwd,
which breaks down if tmux is in use.

And while I was figuring this all out, I realized I could just
hijack tmux's copy-pipe command and do it here!

This is a first take and I expect to iterate a bit, but it's
already quite cool :)