tmux: Don't clear selection on activating plumb(1)
Combined with the previous hotkeys, this makes it easy to keep
my place in a list of things to sort through. For example, say
I want to sift through 100 hits of `git grep -n`. I don't want
to have to spam `n` every time I plumb! Keeping that split open
and the selection in place essentially means I can replace
vim-fugitive's :Ggrep with this :O

Not that I'm definitely gonna do that--fugitive is awesome. I just
could.