tmux: Add support for git-grep under cursor
I do this _all_ the time in vim, so it's super cool to think I
can use tmux + plumb to achieve effectively the same but with any
unstructured text.

As I write this though, I'm realizing that an mless(1) style gless(1)
which has all the rows of git-grep and knows how to feed them to
vim based on custom keybindings and exit codes might be cool.. hm.
Like a git-gui using less :thinking: