bin: Add gg script--git-grep for use in acme(1)
I've been playing around with acme(1) and it really works best to have
commands in the tag that don't have spaces in them. With `gg`, I can
select text, middle click hold on gg and then click the primary button
to pass the selected text as the argument.

This is more convenient than sweeping a tag of `git grep -n`.