theme: Turn on the lights (drop NO_COLOR)
Well, I gave it a good run. I tried forcing monochrome output in most of
my apps. And while it was certainly an interesting experience of
blending everything into "pure text", I found myself maybe the tiniest
bit less effecient in finding things in output / on the screen and it
was hard to justify when pairing with teammates.

That said, going back to regular colors has also been sometimes
overwhelming, which is interesting. So many bright lights! Maybe I'll
tone down the colorscheme.

There was something really cool about "the only color you see is now
more visible". Like I used a subtle pink for syntax errors and it was
reallllly clear when my code was borked.

In the wise words of my friend Cal: "I feel like humans evolved to see
color for a reason, see color, so why not use that to our advantage?"