acme: Add `Note` script, `gv` for current file history
Note is a nice way to just jot down some thoughts in my notes
repo. It starts a new buffer with a default name and a date for the
note, making it easy to edit the name before a Put.

While reviewing a PR in acme, I found I wanted per-file git log history,
which is what gv is! I've also modified gbl to blame the current file,
if no argument is given, which is much easier than having to pass an arg.