acme: Add Bhist/Bls commands to track/navigate Del'd buffers
One of the things I've noticed missing from Vim is the buffer list.
I can :q a file but still get back to it quickly by using :b <file>.

The acme version of this feels like just logging the files as I open
them and then spitting out that log and navigating it with the
plumber!