cleanup: Remove acme / plan9 scripts
I gave acme a real whirl a year or two ago. It's fun
to look back on these and remember how I was trying
to bend it into my workflows... but its just too
impractical for my day-to-day usage (or even hobby
usage). I'll stick to ed(1) when I want to be quirky :)